(Refactor | Green): Move roles/ Module To role/
