+++ role/__init__.py
@@ -1,0 +1,3 @@
+from role.role import Roles
+
+Role = Roles
